+++ griptape/drivers/event_listener/amazon_sqs_event_listener_driver.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
 import json
+from typing import TYPE_CHECKING, Any
 
 from attr import Factory, define, field
 
 from griptape.drivers.event_listener.base_event_listener_driver import BaseEventListenerDriver
-from griptape.events.base_event import BaseEvent
 from griptape.utils import import_optional_dependency
 
 if TYPE_CHECKING:
@@ -19,5 +18,5 @@ class AmazonSqsEventListenerDriver(BaseEventListenerDriver):
     session: boto3.Session = field(default=Factory(lambda: import_optional_dependency("boto3").Session()), kw_only=True)
     sqs_client: Any = field(default=Factory(lambda self: self.session.client("sqs"), takes_self=True))
 
-    def try_publish_event(self, event: BaseEvent) -> None:
-        self.sqs_client.send_message(QueueUrl=self.queue_url, MessageBody=json.dumps({"event": event.to_dict()}))
+    def try_publish_event_payload(self, event_payload: dict) -> None:
+        self.sqs_client.send_message(QueueUrl=self.queue_url, MessageBody=json.dumps(event_payload))