+++ griptape/events/event_listener.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
-from typing import Optional, TYPE_CHECKING
-from attrs import define, field
+from typing import Optional, TYPE_CHECKING, Callable
+from attrs import define, field, Factory
 from .base_event import BaseEvent
 
 if TYPE_CHECKING:
@@ -9,6 +9,7 @@ if TYPE_CHECKING:
 
 @define
 class EventListener:
+    handler: Callable[[BaseEvent], Optional[dict]] = field(default=Factory(lambda: lambda event: event.to_dict()))
     event_types: Optional[list[type[BaseEvent]]] = field(default=None, kw_only=True)
     driver: Optional[BaseEventListenerDriver] = field(default=None, kw_only=True)
 
@@ -16,5 +17,9 @@ class EventListener:
         event_types = self.event_types
 
         if event_types is None or type(event) in event_types:
+            event_payload = self.handler(event)
             if self.driver is not None:
-                self.driver.publish_event(event)
+                if event_payload is not None and isinstance(event_payload, dict):
+                    self.driver.publish_event(event_payload)
+                else:
+                    self.driver.publish_event(event)